Firmware update webUI: -accepts the DNS-325 -only accepts firmware made for the box it is running on

+++ customroot/usr/www/cgi-bin/firmware_proc.cgi
@@ -3,11 +3,16 @@
 . common.sh
 check_cookie
 
-upfile=$(upload_file)
+#debug
+
+if ! upfile=$(upload_file); then
+	msg "Uploading failed: $upfile"
+	exit 0
+fi
 
 if ! test -s $upfile; then
 	rm -f $upfile
-	msg "File is empty."
+	msg "The uploaded file is empty."
 	exit 0
 fi
 
@@ -17,29 +22,27 @@ st=$?
 
 rm -f $upfile
 
+supported="D-Link DNS-320, D-Link DNS-321, D-Link DNS-323, D-Link DNS-325, Conceptronic CH3SNAS, Fujitsu-Siemens DUO 35-LR"
+
 if test $st != "0"; then
-	rm -f kernel initramfs defaults
-	msg "Does not seems to be a legitime firmware file."
+	rm -f kernel initramfs sqimage defaults
+	msg "The uploaded file does not seems to be a legitimate firmware file. The error message was:\n\n\
+$res\n\n\
+Only firmware for the $supported is supported."
 	exit 0
 fi
 
 # The dns323-fw program output identifies the firmware file, evaluate it
 # Model	product_id	custom_id	model_id	sub_id	NewVersion type
-# DNS321		a		1			1			2		1		1
 # DNS323		7		1			1			1		4		0
 # CH3SNAS		7		2			1			1		4		0
-# DUO 35-LR		7		3			1			1		4		0
+# DUO35-LR		7		3			1			1		4		0
+# DNS321		a		1			1			1/2		0/1		1
+# DNS343		9		1			1			1/2		1		2
+# DNS325		0		8			5			1/2		0		3
+# DNS320		0		8			7			1/2		0		4
 # Alt-F-0.1B	1		2			3			4		5		0
 # Alt-F-0.1RC	7		1			1			1		4		0
-
-# Board/Firmware known compatibility.
-# It is not clear cut, as the Alt-F firmware has a DNS323 signature
-#
-#    DNS321 DNS323 CH3SNAS DUO35LR
-# A1  X       V      ?      ?
-# B1  X       V      V      V
-# C1  X       V      ?      ?
-# D1  V       V      X      X
 
 for i in $res; do
 	if echo $i | grep -q ';'; then
@@ -50,31 +53,44 @@ done
 sig=${product_id}${custom_id}${model_id}${sub_id}
 brd=$(cat /tmp/board)
 
-case $sig in
-	"a112") ftype="DNS-321" ;;
+case "$sig" in
+	"0851"|"0852") ftype="DNS-325" ;;
+	"0871"|"0872") ftype="DNS-320" ;;
+	"a111"|"a112") ftype="DNS-321" ;;
+	"9111"|"9112") ftype="DNS-343" ;;
 	"7111") ftype="DNS-323" ;;
 	"7211") ftype="CH3SNAS" ;;
 	"7311") ftype="DUO35LR" ;;
-	"1234") ftype="Alt-F-0.1Bx" ;;
-	*) 
-		rm -f kernel initramfs defaults
-		msg "This firmware file has signature $sig and does not seems to be compatible with the\nD-Link DNS-321, D-Link DNS-323, Conceptronic CH3SNAS or Fujitsu-Siemens DUO 35-LR."
+	*) rm -f kernel initramfs sqimage defaults
+		msg "This firmware file has signature \"$sig\" and does not seems to be compatible with supported firmware.\n\nOnly firmware for the $supported is supported."
 		exit 0
 		;;
 esac
 
-if test "$brd" = "D1" -a "$ftype" != "DNS-321" -a "$ftype" != "DNS-323"; then
-	rm -f kernel initramfs defaults
-	msg "Your box is a DNS-321 and this firmware file was not designed for it".
-elif test "$ftype" = "DNS-321" -a $brd != "D1"; then
-	rm -f kernel initramfs defaults
-	msg "This firmware file was designed for a DNS-321 and your box is not a DNS-321.".
+nomsg="Your box is a $brd and this firmware is for the $ftype"
+
+case "$brd" in
+	"DNS-325-A1") if test $ftype != "DNS-325"; then notcomp=yes; fi ;;
+	"DNS-320-A1") if test $ftype != "DNS-320"; then notcomp=yes; fi ;;
+	"DNS-321-A1") if test $ftype != "DNS-321"; then notcomp=yes; fi ;;
+	"DNS-323-A1"|"DNS-323-B1"|"DNS-323-C1")
+		if test $ftype != "DNS-323" -a $ftype != "CH3SNAS" -a $ftype != "DUO35LR"; then notcomp=yes; fi
+		;;
+	*)	notcomp=yes
+		nomsg="Your box is a $brd and it is not supported.\n\nOnly the $supported boxes are supported."
+		;;
+esac
+
+if test -n "$notcomp"; then
+	rm -f kernel initramfs sqimage defaults
+	msg "$nomsg"
+	exit 0
 fi
 
 html_header
-echo "<h2><center>Firmware Updater</center></h2><pre>"
+echo "<h2><center>Firmware Updater</center></h2>"
 
-echo "$res</pre>"
+#echo "<pre>$res</pre>"
 
 if ! test -s defaults; then
 	flashvendor="disabled"
@@ -86,13 +102,20 @@ else
 	flashfile="checked"
 fi
 
-echo "<h4>Everything looks OK. The firmware file is for a $ftype and you have a rev-$brd board.</h4>"
+if echo $brd | grep -qE "DNS-325|DNS-320"; then
+	try_dis="disabled"
+	recover_dis="disabled"
+fi
 
-if losetup  | grep -q /dev/mtdblock3; then
+echo "<h4>Everything looks OK.<br>$nomsg.</h4>"
+
+if losetup | grep -q /dev/mtdblock3; then
 	runfromflash="disabled"
-	flmsg="<h4><font color=blue>Alt-F is running from flash memory,
-you have to reboot into a special mode before being able to flash.<br>You can however TryIt first.</font></h4>
-<input type=submit name=flash value=\"SpecialReboot\">"
+	flmsg='<h4 class="blue">Alt-F is running from flash memory,
+you have to reboot into a special mode by hitting the "SpecialReboot" button
+and after rebooting run again the "Firmware Updater".
+<br>Alt-F firmware can be tried without flashing it, by hitting the "TryIt" button.</h4>
+<input type=submit name=flash value="SpecialReboot">'
 fi
 
 cat<<-EOF
@@ -107,7 +130,7 @@ cat<<-EOF
 	<p><input type="radio" $flashfile $flashvendor name=flash_defaults value="flashfile">
 	Erase all flashed settings and flash new settings from defaults file (factory settings, the box IP might change)</p>
 
-	<p><input type="radio" $flashvendor name=flash_defaults value="recover">
+	<p><input type="radio" $recover_dis $flashvendor name=flash_defaults value="recover">
 	Erase all flashed settings and recover the last flash settings from vendors backup (the box IP might change)</p>
 $flmsg
 	<input type="submit" $runfromflash name=flash value="FlashIt" onclick="return confirm('\
@@ -117,7 +140,7 @@ There exists the remote possibility that this operation fails, turning your box 
 YOUR HAVE BEEN WARNED.\n\n\
 Continue?')">
 
-	<input type="submit" name=flash value="TryIt" onclick="return confirm('\
+	<input type="submit" $try_dis name=flash value="TryIt" onclick="return confirm('\
 This action only executes the new firmware, it will not flash or change anything else.\n\n\
 While testing the new firmware YOU SHOULD NOT SAVE ANY SETTINGS,\n\
 because they might not be recognized by the current firmware when it runs again.\n\n\
